x.c: modify xsettitle to work with musl libc
Before this commit, running st on a system with musl libc would
result in a segmentation fault. This is because musl implements
locales differently, which means the function
Xutf8TextListToTextProperty would fail, because the locale isn't
supported, leading to nonexistent memory being freed.

I've fixed this by checking if the locale is supported and using
XStringListToTextProperty otherwise.